requirements.txt
Django==4.2.2
Pillow==9.5.0
whitenoise==6.4.0
django-simple-history==3.3.0
djangorestframework==3.14.0
django-environ==0.10.0
django-cors-headers==3.14.0
whitenoise==6.4.0
django-simple-history==3.3.0
Pillow==9.5.0
django-filter==23.2